requirements.txt
aiosmtplib
colorama
loguru
python-decouple
pyotp
qrcode
jinja2
PyJWT